doing some updates in all delete methods in service classes

+++ src/main/java/com/example/pet_care_api/service/impl/BreedServiceImpl.java
@@ -60,7 +60,10 @@ public class BreedServiceImpl implements BreedService {
     }
 
     @Override
-    public void deleteBreedById(Long id) {
+    public void deleteBreedById(Long id) throws BreedNotFoundException {
+        if (!breedRepository.existsById(id)) {
+            throw new BreedNotFoundException("Breed with ID " + id + " not found.");
+        }
         breedRepository.deleteById(id);
     }
